Change Utilization function from float to double to get more precision

+++ include/processor.h
@@ -7,7 +7,7 @@
 
 class Processor {
  public:
-  float Utilization();
+  double Utilization();
 
  private:
   long prevJiffies_ = LinuxParser::Jiffies();
