+++ src/processor.cpp
@@ -7,12 +7,12 @@ using std::vector;
 using std::string;
 
 // Return the aggregate CPU utilization
-float Processor::Utilization() {
+double Processor::Utilization() {
   long jiffies = LinuxParser::Jiffies();
   long activeJiffies = LinuxParser::ActiveJiffies();
   // Calculate difference of previous and current jiffies and activeJiffies
-  float jiffiesDiff = (float)(jiffies - prevJiffies_);
-  float activeJiffiesDiff = (float)(activeJiffies - prevActiveJiffies_);
+  double jiffiesDiff = (jiffies - prevJiffies_);
+  double activeJiffiesDiff = (activeJiffies - prevActiveJiffies_);
   // Update prevJiffies and prevActiveJiffies
   prevJiffies_ = jiffies;
   prevActiveJiffies_ = activeJiffies;
